scraping/requirements.txt
# Financial Data Scraper Requirements

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Financial data sources
yfinance>=0.2.0
fredapi>=0.5.0

# Environment management
python-dotenv>=0.19.0
